Essence Font files divided out a bit
