+++ analysis/monkmistweaver/src/CombatLogParser.ts
@@ -25,7 +25,6 @@ import Buffs from './modules/features/Buffs';
 import Checklist from './modules/features/Checklist/Module';
 import CooldownThroughputTracker from './modules/features/CooldownThroughputTracker';
 import EssenceFontHealingBreakdown from './modules/features/EssenceFontHealingBreakdown';
-import EssenceFontMastery from './modules/features/EssenceFontMastery';
 import EvmVivCastRatio from './modules/features/EvmVivCastRatio';
 import MasteryStats from './modules/features/MasteryStats';
 import MistweaverHealingEfficiencyDetails from './modules/features/MistweaverHealingEfficiencyDetails';
@@ -41,6 +40,8 @@ import CloudedFocus from './modules/shadowlands/legendaries/CloudedFocus';
 import EnvelopingBreath from './modules/spells/EnvelopingBreath';
 import EnvelopingMists from './modules/spells/EnvelopingMists';
 import EssenceFont from './modules/spells/EssenceFont';
+import EssenceFontTargetsHit from './modules/spells/EssenceFontTargetsHit';
+import EssenceFontUniqueTargets from './modules/spells/EssenceFontUniqueTargets';
 import ExpelHarm from './modules/spells/ExpelHarm';
 import InvokeYulon from './modules/spells/InvokeYulon';
 import LifeCocoon from './modules/spells/LifeCocoon';
@@ -89,7 +90,6 @@ class CombatLogParser extends CoreCombatLogParser {
     alwaysBeCasting: AlwaysBeCasting,
     abilities: Abilities,
     cooldownThroughputTracker: CooldownThroughputTracker,
-    essenceFontMastery: EssenceFontMastery,
     checklist: Checklist,
     evmVivCastRatio: EvmVivCastRatio,
     masteryStats: MasteryStats,
@@ -112,6 +112,9 @@ class CombatLogParser extends CoreCombatLogParser {
     envelopingBreath: EnvelopingBreath,
     revival: Revival,
     risingSunKick: RisingSunKick,
+
+    essenceFontUniqueTargets: EssenceFontUniqueTargets,
+    essenceFontTargetsHit: EssenceFontTargetsHit,
 
     // Talents
     chiBurst: ChiBurst,
