+++ analysis/monkmistweaver/src/modules/features/Checklist/Component.tsx
@@ -236,22 +236,6 @@ const MistweaverMonkChecklist = ({ combatant, castEfficiency, thresholds }: Chec
           }
           thresholds={thresholds.thunderFocusTea}
         />
-        <Requirement
-          name={
-            <>
-              <SpellLink id={SPELLS.ESSENCE_FONT.id} /> HOTS Used per Cast
-            </>
-          }
-          thresholds={thresholds.essenceFontMastery}
-        />
-        <Requirement
-          name={
-            <>
-              <SpellLink id={SPELLS.ESSENCE_FONT.id} /> Cancelled Casts
-            </>
-          }
-          thresholds={thresholds.essenceFontCancel}
-        />
       </Rule>
 
       <Rule
