+++ analysis/monkmistweaver/src/modules/features/EssenceFontHealingBreakdown.tsx
@@ -8,15 +8,12 @@ import { STATISTIC_ORDER } from 'parser/ui/StatisticsListBox';
 import React from 'react';
 
 import EssenceFont from '../spells/EssenceFont';
-import EssenceFontMastery from './EssenceFontMastery';
 
 class EssenceFontHealingBreakdown extends Analyzer {
   static dependencies = {
-    essenceFontMastery: EssenceFontMastery,
     essenceFont: EssenceFont,
   };
 
-  protected essenceFontMastery!: EssenceFontMastery;
   protected essenceFont!: EssenceFont;
 
   renderEssenceFontChart() {
@@ -25,24 +22,22 @@ class EssenceFontHealingBreakdown extends Analyzer {
         color: '#00bbcc',
         label: 'Bolt',
         spellId: SPELLS.ESSENCE_FONT.id,
-        value: this.essenceFont.totalHealing + this.essenceFont.totalAbsorbs,
-        valueTooltip: formatThousands(
-          this.essenceFont.totalHealing + this.essenceFont.totalAbsorbs,
-        ),
+        value: this.essenceFont.boltHealing,
+        valueTooltip: formatThousands(this.essenceFont.boltHealing),
       },
       {
         color: '#f37735',
         label: 'Hot',
         spellId: SPELLS.ESSENCE_FONT_BUFF.id,
-        value: this.essenceFont.efHotHeal,
-        valueTooltip: formatThousands(this.essenceFont.efHotHeal),
+        value: this.essenceFont.hotHealing,
+        valueTooltip: formatThousands(this.essenceFont.hotHealing),
       },
       {
         color: '#00b159',
         label: 'Mastery',
         spellId: SPELLS.GUSTS_OF_MISTS.id,
-        value: this.essenceFontMastery.healing,
-        valueTooltip: formatThousands(this.essenceFontMastery.healing),
+        value: this.essenceFont.gomHealing,
+        valueTooltip: formatThousands(this.essenceFont.gomHealing),
       },
     ];
