+++ analysis/monkmistweaver/src/modules/features/MistweaverHealingEfficiencyTracker.tsx
@@ -16,7 +16,6 @@ import RenewingMist from '../spells/RenewingMist';
 import SoothingMist from '../spells/SoothingMist';
 import Vivify from '../spells/Vivify';
 import RefreshingJadeWind from '../talents/RefreshingJadeWind';
-import EssenceFontMastery from './EssenceFontMastery';
 
 class MistweaverHealingEfficiencyTracker extends HealingEfficiencyTracker {
   static dependencies = {
@@ -29,7 +28,6 @@ class MistweaverHealingEfficiencyTracker extends HealingEfficiencyTracker {
     // Custom dependencies
     abilities: Abilities,
     essenceFont: EssenceFont,
-    essenceFontMastery: EssenceFontMastery,
     envelopingMists: EnvelopingMists,
     soothingMist: SoothingMist,
     renewingMist: RenewingMist,
@@ -45,7 +43,6 @@ class MistweaverHealingEfficiencyTracker extends HealingEfficiencyTracker {
   protected castEfficiency!: CastEfficiency;
   protected abilities!: Abilities;
   protected essenceFont!: EssenceFont;
-  protected essenceFontMastery!: EssenceFontMastery;
   protected envelopingMists!: EnvelopingMists;
   protected soothingMist!: SoothingMist;
   protected renewingMist!: RenewingMist;
@@ -100,9 +97,7 @@ class MistweaverHealingEfficiencyTracker extends HealingEfficiencyTracker {
 
   getEssenceFontDetails(spellInfo: SpellInfoDetails) {
     spellInfo.healingDone =
-      this.essenceFont.totalHealing +
-      this.essenceFontMastery.healing +
-      this.essenceFont.totalAbsorbs;
+      this.essenceFont.totalHealing + this.essenceFont.gomEFHealing + this.essenceFont.totalAbsorbs;
     spellInfo.overhealingDone = this.essenceFont.totalOverhealing;
     spellInfo.healingHits = this.essenceFont.targetsEF;
     return spellInfo;
